+++ src/lib/swagger-express-ts/api-operation-post.decorator.ts
@@ -6,6 +6,6 @@ export interface IApiOperationPostArgs extends IApiOperationArgsBase {
 
 export function ApiOperationPost( args: IApiOperationPostArgs ): MethodDecorator {
     return function ( target: any, propertyKey: string | symbol, descriptor: PropertyDescriptor ) {
-        SwaggerService.addOperationPost( args, target, propertyKey );
+        SwaggerService.getInstance().addOperationPost( args, target, propertyKey );
     };
 }