+++ src/lib/swagger-express-ts/express.configurator.ts
@@ -2,6 +2,7 @@ import { Router, Request, Response, NextFunction } from "express";
 import { SwaggerService } from "./swagger.service";
 import * as assert from "assert";
 import { build, ISwaggerBuildDefinition } from "./swagger.builder";
+import { ISwagger } from "./i-swagger";
 
 export interface ISwaggerExpressOptions {
     /**
@@ -18,7 +19,7 @@ export interface ISwaggerExpressOptions {
 
 export function express( options?: ISwaggerExpressOptions ): Router {
     let path: string = "/api-docs/swagger.json";
-    if(options){
+    if ( options ) {
         assert.ok( options.definition, "Definition is required." );
         if ( options.path ) {
             path = options.path;
@@ -34,7 +35,8 @@ export function express( options?: ISwaggerExpressOptions ): Router {
 function buildRouter( path: string ): Router {
     const router: Router = Router();
     router.get( path, ( request: Request, response: Response, next: NextFunction ) => {
-        response.json( SwaggerService.getData() );
+        let data: ISwagger = SwaggerService.getInstance().getData();
+        response.json( data );
     } );
     return router;
 }