+++ src/lib/swagger-specification/index.js
@@ -1,8 +1,0 @@
-"use strict";
-exports.__esModule = true;
-var api_path_decorator_1 = require("api-path.decorator");
-exports.ApiPath = api_path_decorator_1.ApiPath;
-var api_get_decorator_1 = require("api-get.decorator");
-exports.IApiGetArgs = api_get_decorator_1.IApiGetArgs;
-exports.ApiGet = api_get_decorator_1.ApiGet;
-var swaggerJson = {};