+++ src/version/version.controller.spec.js
@@ -1,9 +1,0 @@
-"use strict";
-exports.__esModule = true;
-var chai = require("chai");
-var expect = chai.expect;
-describe("VersionController", function () {
-    it("get", function () {
-        expect(true).is["true"];
-    });
-});