+++ CHANGELOG.md
@@ -1,3 +1,48 @@
+<a name="1.0.0-alpha.5"></a>
+# [1.0.0-alpha.5] (2018-01-29)
+
+## Features
+
+### externalDocs
+
+Example:
+
+```ts
+app.use( swagger.express(
+        {
+            definition : {
+                setInfo : {
+                    title : "Mon api",
+                    version : "1.0"
+                },
+                models : {
+                    Version : {
+                        properties : {
+                            id : {
+                                type : SwaggerDefinitionConstant.Definition.Property.Type.STRING,
+                                required : true
+                            },
+                            name : {
+                                type : SwaggerDefinitionConstant.Definition.Property.Type.STRING,
+                                required : true
+                            },
+                            description : {
+                                type : SwaggerDefinitionConstant.Definition.Property.Type.STRING
+                            },
+                            version : {
+                                type : SwaggerDefinitionConstant.Definition.Property.Type.STRING
+                            }
+                        }
+                    }
+                },
+                externalDocs: {
+                    url: "Mon url"
+                }
+            }
+        }
+    ) );
+```
+
 <a name="1.0.0-alpha.4"></a>
 # [1.0.0-alpha.4] (2018-01-19)
 
@@ -31,7 +76,7 @@ Example:
 app.use( swagger.express(
     {
         definition : {
-            info : {
+            setInfo : {
                 title : "My api",
                 version : "1.0"
             },